+++ packages/visual-stack-docs/src/containers/Components/Docs/loading-animation.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { Panel, Body, Header } from '@cjdev/visual-stack/lib/components/Panel';
 import { Demo, Snippet } from '../../../components/Demo';
 /* s1:start */
-import LoadingAnimation from "@cjdev/visual-stack/lib/components/LoadingAnimation";
+import LoadingAnimation from '@cjdev/visual-stack/lib/components/LoadingAnimation';
 /* s1:end */
 
 export default () => (
@@ -14,7 +14,7 @@ export default () => (
             <Header>Loading Animation</Header>
             <Body>
               {/* s2:start */}
-              <LoadingAnimation loadingMessage={"Loading Data..."}/>
+              <LoadingAnimation loadingMessage={'Loading Data...'} />
               {/* s2:end */}
               <Snippet tag="s1" src={snippets} />
               <Snippet tag="s2" src={snippets} />