+++ packages/visual-stack/src/components/Alert/index.js
@@ -47,9 +47,13 @@ const makeAlert = () => ({
 
   return (
     <div className={classes.join(' ')} {...otherProps}>
-      <div className="vs-alert-icon-container">{iconMap[type] || iconMap.info}</div>
+      <div className="vs-alert-icon-container">
+        {iconMap[type] || iconMap.info}
+      </div>
       <div className="vs-alert-content">{children}</div>
-      <div className="vs-alert-close-container">{closeButtonMap[closeType]}</div>
+      <div className="vs-alert-close-container">
+        {closeButtonMap[closeType]}
+      </div>
     </div>
   );
 };