+++ packages/visual-stack/src/components/BlankSlate/tests/index.test.js
@@ -84,9 +84,9 @@ describe('BlankSlate', () => {
   test('should render', () => {
     const wrapper = mount(<BlankSlate title="title" />);
     expect(wrapper.find(BlankSlate).length).toEqual(1);
-    expect(wrapper.find(".vs-bs-content").length).toEqual(1);
-    expect(wrapper.find(".vs-bs-leftalign").length).toEqual(0);
-    expect(wrapper.find(".vs-bs-title").length).toEqual(1);
+    expect(wrapper.find('.vs-bs-content').length).toEqual(1);
+    expect(wrapper.find('.vs-bs-leftalign').length).toEqual(0);
+    expect(wrapper.find('.vs-bs-title').length).toEqual(1);
     // TODO: Figure out why this isn't working ...
     // expect(wrapper.find(<div className="vs-bs-content"><h1 className="vs-bs-title">"You have no content."</h1></div>).length).toEqual(1);
   });
@@ -94,9 +94,9 @@ describe('BlankSlate', () => {
   test('should render left aligned', () => {
     const wrapper = mount(<BlankSlate title="title" alignment="left-side" />);
     expect(wrapper.find(BlankSlate).length).toEqual(1);
-    expect(wrapper.find(".vs-bs-content").length).toEqual(1);
-    expect(wrapper.find(".vs-bs-leftalign").length).toEqual(1);
-    expect(wrapper.find(".vs-bs-title").length).toEqual(1);
+    expect(wrapper.find('.vs-bs-content').length).toEqual(1);
+    expect(wrapper.find('.vs-bs-leftalign').length).toEqual(1);
+    expect(wrapper.find('.vs-bs-title').length).toEqual(1);
     // TODO: Figure out why this isn't working ...
     // expect(wrapper.find(<div className="vs-bs-content"><h1 className="vs-bs-title">"You have no content."</h1></div>).length).toEqual(1);
   });