+++ packages/visual-stack/src/components/LoadingAnimation/index.js
@@ -2,12 +2,12 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import './styles.css';
 
-const LoadingAnimation = ({loadingMessage = "Loading..."}) => (
+const LoadingAnimation = ({ loadingMessage = 'Loading...' }) => (
   <div className="loading-animation loading-container">
     <div className="loading-animation loading-bars">
-      <div/>
-      <div/>
-      <div/>
+      <div />
+      <div />
+      <div />
     </div>
     <div className="loading-animation loading-text">{loadingMessage}</div>
   </div>